Increase acceptable length of custom fields and allow null values
Currently the maximum length of these custom fields is 25 characters.  It's not too difficult to imagine a case in which this is far too little.  Also, currently none of the fields are allowed to remain null, which for most uses doesn't make sense as every item wont necessarily have all the fields populated.